fix: extract error utilities
