+++ packages/javascript/tests/utils/error.test.ts
@@ -1,11 +1,11 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
-import { getErrorFromEvent } from '../../src/utils/event';
+import { getErrorFromErrorEvent } from '../../src/utils/error';
 
 vi.mock('@hawk.so/core', () => ({ log: vi.fn(), isLoggerSet: vi.fn(() => true), setLogger: vi.fn() }));
 
 import Sanitizer from '../../src/modules/sanitizer';
 
-describe('getErrorFromEvent', () => {
+describe('getErrorFromErrorEvent', () => {
   beforeEach(() => {
     vi.clearAllMocks();
   });
@@ -15,7 +15,7 @@ describe('getErrorFromEvent', () => {
       const error = new Error('Test error');
       const event = new ErrorEvent('error', { error });
 
-      const result = getErrorFromEvent(event);
+      const result = getErrorFromErrorEvent(event);
 
       expect(result).toBe(error);
     });
@@ -24,7 +24,7 @@ describe('getErrorFromEvent', () => {
       const error = new DOMException('Network error', 'NetworkError');
       const event = new ErrorEvent('error', { error });
 
-      const result = getErrorFromEvent(event);
+      const result = getErrorFromErrorEvent(event);
 
       expect(result).toBe('<instance of DOMException>');
     });
@@ -32,7 +32,7 @@ describe('getErrorFromEvent', () => {
     it('should return the message when event.error is not provided and message is a string', () => {
       const event = new ErrorEvent('error', { message: 'Script error.' });
 
-      const result = getErrorFromEvent(event);
+      const result = getErrorFromErrorEvent(event);
 
       expect(result).toBe('Script error.');
     });
@@ -40,7 +40,7 @@ describe('getErrorFromEvent', () => {
     it('should return empty string when event.error is not provided and message is empty', () => {
       const event = new ErrorEvent('error', { message: '' });
 
-      const result = getErrorFromEvent(event);
+      const result = getErrorFromErrorEvent(event);
 
       expect(result).toBe('');
     });
@@ -51,7 +51,7 @@ describe('getErrorFromEvent', () => {
       const reason = new Error('Promise rejected');
       const event = new PromiseRejectionEvent('unhandledrejection', { promise: Promise.resolve(), reason });
 
-      const result = getErrorFromEvent(event);
+      const result = getErrorFromErrorEvent(event);
 
       expect(result).toBe(reason);
     });
@@ -60,7 +60,7 @@ describe('getErrorFromEvent', () => {
       const reason = 'Promise rejected with string';
       const event = new PromiseRejectionEvent('unhandledrejection', { promise: Promise.resolve(), reason });
 
-      const result = getErrorFromEvent(event);
+      const result = getErrorFromErrorEvent(event);
 
       expect(result).toBe(reason);
     });
@@ -69,7 +69,7 @@ describe('getErrorFromEvent', () => {
       const reason = { code: 'ERR_001', details: 'Something went wrong' };
       const event = new PromiseRejectionEvent('unhandledrejection', { promise: Promise.resolve(), reason });
 
-      const result = getErrorFromEvent(event);
+      const result = getErrorFromErrorEvent(event);
 
       expect(result).toBe('Promise rejected with {"code":"ERR_001","details":"Something went wrong"}');
     });
@@ -77,7 +77,7 @@ describe('getErrorFromEvent', () => {
     it('should return undefined when event.reason is not provided', () => {
       const event = new PromiseRejectionEvent('unhandledrejection', { promise: Promise.resolve(), reason: undefined });
 
-      const result = getErrorFromEvent(event);
+      const result = getErrorFromErrorEvent(event);
 
       expect(result).toBeUndefined();
     });
@@ -85,7 +85,7 @@ describe('getErrorFromEvent', () => {
     it('should return null when event.reason is null', () => {
       const event = new PromiseRejectionEvent('unhandledrejection', { promise: Promise.resolve(), reason: null });
 
-      const result = getErrorFromEvent(event);
+      const result = getErrorFromErrorEvent(event);
 
       expect(result).toBeNull();
     });
@@ -96,7 +96,7 @@ describe('getErrorFromEvent', () => {
 
       const event = new PromiseRejectionEvent('unhandledrejection', { promise: Promise.resolve(), reason: circularObj });
 
-      const result = getErrorFromEvent(event);
+      const result = getErrorFromErrorEvent(event);
 
       expect(result).toContain('Promise rejected with');
       expect(result).toContain('<circular>');
